feat(OUT-1313): add new task notifications service
